FILE: moves selectRandom
