+++ libs/spirit-islander/util/src/lib/app/select-random.ts
