+++ libs/spirit-islander/util/src/lib/functions/select-boards.ts
@@ -2,7 +2,7 @@ import { BOARDS, Board, BalancedBoardName } from '../game/boards';
 import type { MapName } from '../game/maps';
 import type { Players } from '../game/players';
 import { getOptionsByName } from '../game/get-options';
-import { selectRandom } from './internal/select-random';
+import { selectRandom } from '../app/select-random';
 
 export function selectBoards(
   mapName: MapName,
